sync collate.SetNewCollationEnabledForTest
